+++ src/action/authorize/paymentMethod/any.ts
@@ -1,6 +1,6 @@
 import * as ActionFactory from '../../../action';
 import * as CheckMovieTicketActionFactory from '../../../action/check/paymentMethod/movieTicket';
-import { IMovieTicket } from '../../../action/trade/pay';
+import { AvailablePaymentMethodType, IMovieTicket } from '../../../action/trade/pay';
 import { ActionType } from '../../../actionType';
 import * as PayTransactionFactory from '../../../assetTransaction/pay';
 import { AssetTransactionType } from '../../../assetTransactionType';
@@ -11,18 +11,12 @@ import { IPropertyValue } from '../../../propertyValue';
 import { TransactionType } from '../../../transactionType';
 import * as AuthorizeActionFactory from '../../authorize';
 
-// 最適化(2022-06-01~)
 export type IAgent = ActionFactory.IParticipantAsWebApplication | ActionFactory.IParticipantAsPerson;
 export type IRecipient = ActionFactory.IParticipantAsSeller;
 
 export enum ResultType {
     Payment = 'Payment'
 }
-
-/**
- * 汎用決済方法タイプ
- */
-export type IAnyPaymentMethod = string;
 
 /**
  * 進行中取引
@@ -75,7 +69,7 @@ export interface IObject {
     /**
      * 決済方法
      */
-    paymentMethod: IAnyPaymentMethod;
+    paymentMethod: AvailablePaymentMethodType;
     /**
      * 決済ID
      */
@@ -112,20 +106,31 @@ export interface IObject {
      */
     movieTickets?: IMovieTicket[];
 }
+export interface IResultPaymentMethod {
+    /**
+     * 決済方法区分
+     */
+    typeOf: AvailablePaymentMethodType;
+    amount?: {
+        /**
+         * 決済カード通貨区分
+         */
+        currency?: string;
+    };
+}
 export interface IResult {
     /**
      * The identifier for the account the payment will be applied to.
      */
     accountId: string;
-    /**
-     * The amount of money.
-     * 廃止(2023-08-07~)
-     */
-    // amount: number;
+    // amount: number; // 廃止(2023-08-07~)
+    // paymentMethod?: AvailablePaymentMethodType; // 廃止 Use paymentMethodAsObject(2023-08-18~)
     /**
      * 決済方法
+     * amount.currencyに対応するために追加(2023-08-13~)
+     * startDate>=2022-08-16T00:00:00Zのアクションの関しては互換性維持済
      */
-    paymentMethod: IAnyPaymentMethod;
+    paymentMethodAsObject: IResultPaymentMethod;
     /**
      * 決済ID
      */
@@ -183,7 +188,3 @@ export interface IAttributes extends AuthorizeActionFactory.IAttributes<IObject,
  * 決済承認アクション
  */
 export type IAction = ActionFactory.IAction<IAttributes>;
-
-export {
-    ITotalPaymentDue
-};